% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_classifier.r
\name{binary_classifier}
\alias{binary_classifier}
\title{This function trains a binary classifier using LASSO for feature selection and elastic net for parameter tuning.}
\usage{
binary_classifier(
  y,
  X,
  nrepeats = 10,
  nfolds = 5,
  final_fold = NA,
  seed = 1337
)
}
\arguments{
\item{y}{A factor, numeric, or integer vector of binary outcomes (0/1 or two levels).}

\item{X}{A data frame or matrix of predictor variables.}

\item{nrepeats}{An integer specifying the number of times to repeat the LASSO feature selection process. Default is 10.}

\item{nfolds}{An integer specifying the number of folds for cross-validation. Default is 5.}

\item{final_fold}{A logical vector indicating the final training set. If NA, it will be generated using caret's createFolds function. Default is NA.}

\item{seed}{An integer specifying the random seed for reproducibility. Default is 1337.}
}
\value{
A list containing the final model, selected features, optimal alpha and lambda values, and the training set used for the final model.
}
\description{
This function trains a binary classifier using LASSO for feature selection and elastic net for parameter tuning.
}
