% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biotype_colors.R
\docType{data}
\name{biotype_colors}
\alias{biotype_colors}
\title{Biotype colors for pretty plotting (use with summarize_biotypes())}
\format{
An object of class \code{list} of length 13.
}
\usage{
biotype_colors
}
\description{
Biotype colors for pretty plotting (use with summarize_biotypes())
}
\keyword{datasets}
