% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{crate_client}
\alias{crate_client}
\alias{has_crate}
\title{Create a crate client}
\usage{
crate_client(url = NULL)

has_crate(url = NULL)
}
\arguments{
\item{url}{The URL of the server to connect to.  By default this
will use \code{http://127.0.0.1:4200}.  You will need to include
both the protocol and the port number in most cases.}
}
\description{
Create a crate client.  The function \code{has_crate} is a small
convenience funtion to test if there is a crate server running at
a given url.
}
\author{
Rich FitzJohn
}

