% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crater.R
\docType{package}
\name{crater}
\alias{crater}
\alias{crater-package}
\title{A simple example of Rust code integrated into an R package}
\description{
This is (currently) an example package. Duncan Temple-Lang
wrote a little demonstrator for calling rust code from R: this wraps that in
a package interface and demonstrates how seamless the
result can look to the end user.
}

