% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeqDataSetFromCrate.r
\name{DESeqDataSetFromCrate}
\alias{DESeqDataSetFromCrate}
\title{Make a DDS object from a CREATE H5SummarizedExperiment}
\usage{
DESeqDataSetFromCrate(
  create_se,
  design,
  min_expr = 3,
  min_frac = 0.25,
  min_lib_size = 1e+05
)
}
\arguments{
\item{create_se}{A H5SummarizedExperiment object created by the CRATE pipeline}

\item{design}{A formula specifying the model design}
}
\value{
A DESeqDataSet object
}
\description{
Make a DDS object from a CREATE H5SummarizedExperiment
}
