% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeqDataSetFromCrate.r
\name{DESeqDataSetFromCreate}
\alias{DESeqDataSetFromCreate}
\title{Make a DDS object from a CREATE H5SummarizedExperiment}
\usage{
DESeqDataSetFromCreate(create, design)
}
\arguments{
\item{create}{A H5SummarizedExperiment object created by the CRATE pipeline}

\item{design}{A formula specifying the model design}
}
\value{
A DESeqDataSet object
}
\description{
Make a DDS object from a CREATE H5SummarizedExperiment
}
