% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gmt.r
\name{make_gmt}
\alias{make_gmt}
\title{Load in gene lists as a 'GMT' object}
\usage{
make_gmt(genes, names, save_to = NA)
}
\arguments{
\item{genes}{A list of gene sets (a list of list of character vectors)}

\item{names}{The name of each gene set}

\item{save_to}{Optional path to save the GMT file}
}
\value{
A GMT object
}
\description{
Load in gene lists as a 'GMT' object
}
