% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_enrichment.r, R/plot_ontology.r
\name{plot_enrichment}
\alias{plot_enrichment}
\title{Plot geneset enrichment from CRATE dds}
\usage{
plot_enrichment(
  dds,
  gene_list,
  ontology = "BP",
  contrast = NA,
  pval = 0.01,
  fold_change = 1,
  signif = 0.01,
  top = 10,
  hide_insig = TRUE,
  colors = c("#50FA7B", "#44475A"),
  labeller = NA,
  ...
)

plot_enrichment(
  dds,
  gene_list,
  ontology = "BP",
  contrast = NA,
  pval = 0.01,
  fold_change = 1,
  signif = 0.01,
  top = 10,
  hide_insig = TRUE,
  colors = c("#50FA7B", "#44475A"),
  labeller = NA,
  ...
)
}
\arguments{
\item{dds}{A DESeqDataSet object from \strong{after} running DESeq()}

\item{gene_list}{A list of gene sets for enrichment analysis}

\item{ontology}{The ontology to use}

\item{contrast}{The DESeq2 contrast to use}

\item{pval}{The p-value threshold}

\item{fold_change}{The fold change threshold}

\item{signif}{The significance level}

\item{top}{The number of top pathways to display}

\item{hide_insig}{Whether to hide insignificant pathways}

\item{colors}{A vector of colors for the plot}

\item{labeller}{A function for labeling the pathways}

\item{geneset}{A list of gene sets for enrichment analysis}

\item{signif_level}{The significance level for filtering pathways}
}
\value{
A ggplot2 object

A ggplot2 object
}
\description{
Plot geneset enrichment from CRATE dds

Plot geneset enrichment from CRATE dds
}
