% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Plot Differential expression from DESeq2 results as a volcano plot}
\usage{
plot_volcano(
  dds,
  contrast,
  color_by = "gene_biotype",
  colors = biotype_colors,
  p_threshold = 0.01,
  lfc_threshold = 1,
  use_shrinkage = TRUE,
  base_size = 6
)
}
\arguments{
\item{dds}{A DESeqDataSet object \strong{after} running DESeq()}

\item{contrast}{A character vector of length 3 specifying the contrast to plot, the numerator and the denominator}

\item{color_by}{A column name in the rowData of the dds to color points by. Default is "gene_biotype"}

\item{colors}{A named vector of colors to use for the different levels of color_by. Default is biotype_colors (assume biotypes are summarized)}

\item{p_threshold}{Adjusted p-value threshold for significance. Default is 0.01}

\item{lfc_threshold}{Log2 fold change threshold for significance. Default is 1}

\item{use_shrinkage}{Whether to use lfsShrink with apeglm}

\item{base_size}{Base size for the plot text and points. Default is 8}
}
\value{
A ggplot2 object representing the volcano plot
}
\description{
Plot Differential expression from DESeq2 results as a volcano plot
}
