% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_plot.r
\name{scale_plot}
\alias{scale_plot}
\title{Refits a already generated plot to a new size}
\usage{
scale_plot(plot, owid, ohgt, scale)
}
\arguments{
\item{plot}{The plot object to be refitted}

\item{owid}{The original width of the plot (in)}

\item{ohgt}{The original height of the plot (in)}

\item{scale}{The new width of the plot (in)}
}
\value{
The refitted plot object
}
\description{
Refits a already generated plot to a new size
}
