% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_biotypes.r
\name{summarize_biotypes}
\alias{summarize_biotypes}
\title{Collect the GENCODE/RepeatMasker biotypes into relevant subsets}
\usage{
summarize_biotypes(
  data,
  biotypes = "gene_biotype",
  gene_ids = "gene_id",
  gene_names = "gene_name",
  biotypes_to = NA
)
}
\arguments{
\item{data}{A data frame containing a column with biotype information}

\item{biotypes}{A string specifying the column name with biotype information
\link{default="gene_biotype"}}
}
\value{
The data frame with the summarized biotypes
}
\description{
Collect the GENCODE/RepeatMasker biotypes into relevant subsets
}
