% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_create.r
\name{theme_create}
\alias{theme_create}
\title{A publication-ready theme for ggplot2}
\usage{
theme_create(base_size = 10)
}
\arguments{
\item{base_size}{Base font size (default is 10)}
}
\value{
A ggplot2 theme
}
\description{
A publication-ready theme for ggplot2
}
