% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_results.r
\name{collect_results}
\alias{collect_results}
\title{Collect the results from a DESeq2 run and add in gene biotype information}
\usage{
collect_results(dds, ...)
}
\arguments{
\item{dds}{A DESeqDataSet object from DESeqDataSetFromCreate}

\item{...}{Additional arguments to pass to DESeq2::results}
}
\value{
A data.frame of results with gene biotype information
}
\description{
Collect the results from a DESeq2 run and add in gene biotype information
}
