% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_counts.r
\name{load_counts}
\alias{load_counts}
\title{Make a DDS object from a CREATE H5SummarizedExperiment}
\usage{
load_counts(countsDir)
}
\arguments{
\item{create}{A H5SummarizedExperiment object created by the CREATE pipeline}
}
\value{
A H5-backed SummarizedExperiment object
}
\description{
Make a DDS object from a CREATE H5SummarizedExperiment
}
