% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_composition.r
\name{plot_composition}
\alias{plot_composition}
\title{Make a plot of the composition of each sample}
\usage{
plot_composition(
  dds,
  position = "fill",
  fill = "gene_biotype",
  facet = NA,
  subset = NA,
  nested = FALSE,
  add_counts = FALSE,
  base_size = 10
)
}
\arguments{
\item{dds}{A DESeqDataSet object from DESeqDataSetFromCreate}

\item{position}{Position adjustment for geom_bar \link{default="fill"}}

\item{fill}{Variable to fill bars with \link{default=gene_biotype}}

\item{facet}{Variables to facet the plot with \link{default=NA}}

\item{subset}{Samples to include in the plot \link{default=NA}}
}
\value{
A ggplot2 object
}
\description{
Make a plot of the composition of each sample
}
