% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_expression.r
\name{plot_expression}
\alias{plot_expression}
\title{Plot geneset enrichment from CREATE dds}
\usage{
plot_expression(
  dds,
  gene_ids,
  counts = "normalized",
  contrast = NA,
  color_by = NA,
  show_ns = TRUE,
  step = 0.1,
  nrow = NA,
  size = 8,
  labeller = NA
)
}
\arguments{
\item{dds}{A DESeqDataSet object from DESeqDataSetFromCreate}

\item{gene_ids}{A character vector of gene IDs or a logical vector indicating which genes to plot}

\item{counts}{Either "raw" or "normalized" counts to plot}

\item{contrast}{A vector specifying the contrast for DESeq2 results, or a single factor to plot all levels}

\item{color_by}{The column in colData to color the boxplots by - defaults to gene IDs}

\item{show_ns}{Whether to show non-significant comparisons \link{default = TRUE}}

\item{step}{The step size for positioning significance annotations \link{default = 0.1}}

\item{nrow}{The number of rows in the facet wrap - defaults to a square}

\item{size}{Base font size for the plot}
}
\value{
A ggplot2 object
}
\description{
Plot geneset enrichment from CREATE dds
}
