% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_crate.r
\name{theme_crate}
\alias{theme_crate}
\title{A publication-ready theme for ggplot2}
\usage{
theme_crate(base_size = 6)
}
\arguments{
\item{base_size}{Base font size (default is 10)}
}
\value{
A ggplot2 theme
}
\description{
A publication-ready theme for ggplot2
}
